service/src/main/java/com/moneysupermarket/componentcatalog/service/controllers/SummaryController.java
package com.moneysupermarket.componentcatalog.service.controllers;

import com.moneysupermarket.componentcatalog.sdk.models.GetSummaryResponse;
import com.moneysupermarket.componentcatalog.service.partialresponse.PartialResponse;
import com.moneysupermarket.componentcatalog.service.services.ComponentService;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequiredArgsConstructor
@RequestMapping("/v1/summary")
public class SummaryController {

    private final ComponentService componentService;

    @GetMapping
    @PartialResponse
    public GetSummaryResponse getSummary() {
        return new GetSummaryResponse(componentService.getSummary());
    }
}
